misc/HtmlTestSuite/results/test-0150.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="mailto:mstensho@chromium.org" rel="author" title="Morten Stenshorne"/>
<link href="https://drafts.csswg.org/css-sizing/#available" rel="help"/>
<link href="https://drafts.csswg.org/css-sizing/#percentage-sizing" rel="help"/>
</head><body><p>There should be a green square below, and no red.</p>
<div id="container" style="width:100px; background:red;">
  <div style="float:right; width:50px; height:100px; background:green;"/>
  <div data-expected-width="50" style="display:table; max-width:50%; height:100px; background:green;">
    <div style="float:left; width:20px; height:10px;"/>
    <div style="float:left; width:20px; height:10px;"/>
    <div style="float:left; width:20px; height:10px;"/>
  </div>
</div>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>
<script>
  checkLayout("#container");
</script>
</body></html>
